Fix debouncing
