+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/home/user2504/projects/github-users/src/reportWebVitals.js":"1","/home/user2504/projects/github-users/src/App.js":"2","/home/user2504/projects/github-users/src/hooks/useDebounce.js":"3"},{"size":362,"mtime":1611847566767,"results":"4","hashOfConfig":"5"},{"size":1186,"mtime":1611864244321,"results":"6","hashOfConfig":"5"},{"size":385,"mtime":1611863982216,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"t7ezk7",{"filePath":"10","messages":"11","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/user2504/projects/github-users/src/reportWebVitals.js",[],"/home/user2504/projects/github-users/src/App.js",["14"],"/home/user2504/projects/github-users/src/hooks/useDebounce.js",[],{"ruleId":"15","severity":1,"message":"16","line":21,"column":6,"nodeType":"17","endLine":21,"endColumn":8,"suggestions":"18"},"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'perPage'. Either include it or remove the dependency array.","ArrayExpression",["19"],{"desc":"20","fix":"21"},"Update the dependencies array to be: [perPage]",{"range":"22","text":"23"},[553,555],"[perPage]"]
