+++ src/App.js
@@ -1,6 +1,7 @@
 import "./App.css";
 import { useEffect, useState } from "react";
 import { getUsersListApi } from "./api";
+import {useDebounce} from './hooks/useDebounce'
 
 function App() {
   const [userList, setUserList] = useState([]);
@@ -9,30 +10,7 @@ function App() {
   const [isSearching, setIsSearching] = useState(false);
   const debouncedPerPage = useDebounce(perPage, 1000);
 
-  function useDebounce(value, delay) {
-    const [debouncedValue, setDebouncedValue] = useState(value);
-    useEffect(() => {
-      const handler = setTimeout(() => {
-        setDebouncedValue(value);
-      }, delay);
-      return () => {
-        clearTimeout(handler);
-      };
-    }, [value, delay]);
-    return debouncedValue;
-  }
-
-  useEffect(() => {
-    if (debouncedPerPage) {
-      setIsSearching(true);
-      getUsersListApi(debouncedPerPage).then((userList) => {
-        setIsSearching(false);
-        setUserList(userList);
-      });
-    } else {
-      setUserList([]);
-    }
-  }, [debouncedPerPage]);
+  const SINCE = 1
 
   useEffect(() => {
     const params = {
@@ -42,13 +20,19 @@ function App() {
     getUsersListApi(params).then((res) => setUserList(res));
   }, []);
 
+
   useEffect(() => {
-    const params = {
-      perPage: perPage,
-      since: 1,
-    };
-    getUsersListApi(params).then((res) => setUserList(res));
-  }, [perPage]);
+    if (debouncedPerPage) {
+      setIsSearching(true);
+      
+      getUsersListApi({perPage: debouncedPerPage, since: SINCE }).then((userList) => {
+        setIsSearching(false);
+        setUserList(userList);
+      });
+    } else {
+      setUserList([]);
+    }
+  }, [debouncedPerPage]);
 
   return (
     <div className="App">
@@ -57,8 +41,7 @@ function App() {
         onChange={(e) => setPerPage(e.target.value)}
         value={perPage}
       ></input>
-      {isSearching && <div>Searching ...</div>}
-      {userList.map((el) => (
+      {isSearching ? <div>Searching ...</div> : userList.map((el) => (
         <div>{el.login}</div>
       ))}
     </div>
